requirements.txt
click
requests
cloudinary
fastapi[standard]
pyprojroot
pygwalker
pymongo
pytz
python-multipart
jupyter
ipykernel
python-decouple
psycopg2
psycopg2-binary
SQLAlchemy
colorama
openpyxl
xlrd
numpy
pandas
matplotlib
seaborn
kaleido
plotly
squarify
calplot
plotly-calplot
wordcloud
matplotlib-venn
anywidget
gw-dsl-parser
quickjs
segment-analytics-python
ipywidgets
papermill
rpy2